updates ids for adagrams
also, the actual adagrams repo itself was set to use main as the default branch.
this was breaking Learn's ability to find the upstream.
Explicitly adding tree/main to the upstream _did_ let it find it.
But Since we haven't updated the other projects to main and fully tested the process,
or added documentation to the submission steps,
it seemed better to set the default branch to master for now.